requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
-e ./etebase-server
aiofiles==0.6.0
aioredis==1.3.1
appdirs==1.4.4
apscheduler==3.7.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
asgiref==3.3.1
async-timeout==3.0.1
azure-common==1.1.27
azure-core==1.16.0
azure-storage-blob==12.8.1
azure-storage-common==2.1.0
black==21.6b0
certifi==2021.5.30
cffi==1.14.4
charset-normalizer==2.0.3; python_version >= '3.0'
click==7.1.2
cryptography==3.4.7; python_version >= '3.6'
django-apscheduler==0.6.0
django-cors-headers==3.7.0
django-ninja==0.13.2
django==3.1.4
fastapi==0.63.0
flake8==3.9.2
git+https://github.com/pjsier/django-storages.git@fb519de50b136dbdff37d72e8c564099c0aa088e#egg=django-storages
h11==0.11.0
hiredis==1.1.0
httptools==0.1.1
idna==3.2; python_version >= '3.0'
isodate==0.6.0
loguru==0.5.3
mccabe==0.6.1
msgpack==1.0.2
msrest==0.6.21
mypy-extensions==0.4.3
mypy==0.910
oauthlib==3.1.1; python_version >= '3.6'
pathspec==0.9.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.20
pydantic==1.7.3
pyflakes==2.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pynacl==1.4.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-decouple==3.4
python-dotenv==0.15.0
python-http-client==3.3.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.4
pyyaml==5.3.1
regex==2021.7.6
requests-oauthlib==1.3.0
requests==2.26.0; python_version >= '3.0'
sendgrid==6.7.1
sentry-sdk==1.3.1
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
starlette==0.13.6
stripe==2.60.0
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn[standard]==0.13.2
uvloop==0.14.0
watchgod==0.6
websockets==8.1
